+++ Verizon/MPVerizonInterstitialCustomEvent.m
@@ -218,6 +218,13 @@
         MPLogAdEvent([MPLogEvent adDidDisappearForAdapter:NSStringFromClass(self.class)], [self getAdNetworkId]);
         
         [self.delegate fullscreenAdAdapterAdDidDisappear:self];
+        
+        // Signal that the fullscreen ad is closing and the state should be reset.
+        // `fullscreenAdAdapterAdDidDismiss:` was introduced in MoPub SDK 5.15.0.
+        if ([self.delegate respondsToSelector:@selector(fullscreenAdAdapterAdDidDismiss:)]) {
+            [self.delegate fullscreenAdAdapterAdDidDismiss:self];
+        }
+        
         [self invalidate];
     });
 }