+++ Verizon/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+* 1.8.1.1
+    * Add support for `fullscreenAdAdapterAdDidDismiss:` to signal that the fullscreen ad is closing and the state should be reset. To use this adapter version, you need v5.15.0 of the MoPub iOS SDK at the minimum.
+    * Remove `nativeVideoView` as part of MoPub's native video code removal. This does not impact Verizon. No external changes or actions rerquired for publishers. 
+
 * 1.8.1.0
     * This version of the adapters has been certified with Verizon 1.8.1 and MoPub 5.14.1.